Reducing warning count a bit...
git-svn-id: https://svn.java.net/svn/glazedlists~svn/trunk@2101 696c4f5c-992c-4387-df9e-f6c076136008

+++ extensions/ktable/source/ca/odell/glazedlists/swt/AbstractKTableFormat.java
@@ -34,7 +34,7 @@ public abstract class AbstractKTableFormat implements KTableFormat {
     /** {@inheritDoc} */
     public int getColumnWidth(int column) {
         prepareDefaultColumnWidths();
-        return ((Integer)columnWidths.get(column)).intValue();
+        return columnWidths.get(column).intValue();
     }
 
     /** {@inheritDoc} */
